Fine_tune_LLM/requirements.txt
transformers>=4.35.0
datasets>=2.14.0
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
torch>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
tqdm>=4.66.0
pandas
pyyaml
